Another week of work...

+++ src/OOBase/Buffer.h
@@ -105,7 +105,7 @@ namespace OOBase
 
 		~Buffer();
 
-		AtomicInt<unsigned long> m_refcount; ///< The reference count.
+		AtomicInt<size_t> m_refcount; ///< The reference count.
 
 		size_t	m_capacity;	///< The total allocated bytes for \p m_buffer.
 		char*	m_buffer;	///< The actual underlying buffer.
